Overlay - Added support for cropping and committing updates
Added 2 APIs to overlay_data_device_t to support cropping.
Added the API commitUpdates() to overlay_control_device_t. This API must be called after a setPosition or setParameter call, for the changes to take
effect. In this way, we avoid multiple stream on/off calls and transition is more smooth.

Signed-off-by: Anu Sundararajan <sanuradha@ti.com>