Overlay.h: Added support for resizeInput()
There are cases where Camera HAL will be receiving preview frames from the image pipe that are actually larger than the client requested size. At the
moment the overlay architecture does not allow the data side of the overlay to dictate a preview input frame size. As an interim solution, we have
added a new overlay data side API that would allow Camera HAL to resize the input frame size.

Signed-off-by: Anu Sundararajan <sanuradha@ti.com>