overlay.h - Added new APIs
Added the API commitUpdates() to overlay_control_device_t. This API must be called after a setPosition or setParameter call, for the changes to take
effect. In this way, we avoid multiple stream on/off calls and the transition is more smooth.

Added the API resizeInput() to overlay_data_device_t. There are cases where the Camera HAL will be receiving preview frames from the image pipe that
are actually larger than the client requested size. At the moment, the overlay architecture does not allow the data side of overlay to dictate a
preview input frame size. As an interim solution, we have added a new overlay data side API that would allow the Camera HAL to resize the input frame
size.

Added the API setCrop() to overlay_data_device_t. This API can be used to set the crop parameters.

Signed-off-by: Anu Sundararajan <sanuradha@ti.com>