Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

(4 inline comments)

Thank you for this change. The new definitions look good but I have a few comments and questions.

For now we cannot break backward binary compatibility by changing the size of the sensor_data_t structure. This will be addressed in the future, for now we have to punt on the part of the change.

Patch-set: 1
Label: Code-Review=-1
Label: Verified=0