Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Thanks for the contribution! Unfortunately we can't accept only a constant. We also need to a specification and documentation for the meaning of the data returned for the sensor type (including its units), required minimum frequency of results returned, and so on. For an example of the sort of detail we need, please see the 2.3 Compatibility Definition Document (CDD) at http://source.android.com/compatibility/

Patch-set: 1
Label: Code-Review=-1
Label: Verified=0