Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

the patch is ok, but we solved the issue slighty differently for Honeycomb which is about to be pushed to the AOSP servers.

As such, I would prefer if you didn't submit this to avoid a conflict (which would delay the push).

Thanks for your comprehension.

Patch-set: 2
Reviewer: David Turner <1000411@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1
Label: Verified=0