Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Why did you change the uint32 to int? That seems like it would cause issues to existing code.

Also, the IPv6 address should be a struct in6_addr instead of just a character array? That way, existing functions can be used to parse it and so on.

The same comment applies to IPv4 (struct in_addr instead of uint_32) but it's probably best left as uint32 for compatibility.

Patch-set: 1
Label: Code-Review=-1