Update patch set 2
Patch Set 2:

int is reverted back to unit32_t.  The original thought was that java is signed.  But we probably do not have to change it, since it is already accepted to the baseline.  As for ipv6, we just want to get the address across to Java space. Our understanding is that all what matters is the 16 bytes of address.  It is not so much about a generic data structure, but how most efficient to send the data across and have the object constructed in Java.  This is the thought behind the 16 byte size char array.  We kept it so for now, looking forward to further comments.

Patch-set: 2
Label: Code-Review=0