[VIBRA] Adapt AOSP (default) vibrator HAL into a hardware module.
Android's implementation of vibrator needs to be done inside
a hardware module, so that it can make a vendor implementation
possible. This is the case for Intel vendor implementation.
Hw module's name becomes vibrator.default.so.

Note: This patch is related to a change in platforms/hardware/libhardware_legacy

Change-Id: I7ef481f8a437a35e399b0310be9abc951f482e25
Author: Vincent Becker <vincentx.becker@intel.com>
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>
Signed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Signed-off-by: David Wagner <david.wagner@intel.com>
Author-tracking-BZ: 49760