Update patch set 1
Patch Set 1:

I chose not to make the change in btif_pan.c so that we can maintain forward compatibility between Bluetooth.apk and bluetooth.default.so. Although both should be updated at the same time, there may be developers who push one file but not the other. Instead of silently failing and having engineers waste time debugging the issue, I chose to fix the problem by just relabeling the parameters in the interface to match the parameters actually being passed.

Patch-set: 1