Update patch set 1
Patch Set 1: Code-Review-1

Thanks for the patch, but please state the rationale for this change in the commit message.

I'm putting a -1 here because according to [1], this could result in unwanted behaviour on some devices. I.e. we may want this to be controlled by a platform/board configuration option.

I would also recommend adding some sort of warning if the first ioctl() is not available, to make potential regressions easier to diagnose.

[1] https://groups.google.com/forum/#!topic/android-framework/2-S9ialdeNs

Patch-set: 1
Label: Code-Review=-1