Update patch set 1
Patch Set 1:

> Thanks for the patch, but please state the rationale for this
 > change in the commit message.

Ok. I'll do that in my next patch set.

 > 
 > I'm putting a -1 here because according to [1], this could result
 > in unwanted behaviour on some devices. I.e. we may want this to be
 > controlled by a platform/board configuration option.

I agree with the discussions [1] and [2] that neither API is well documented about their usage and most of the time it really depends on the respective framebuffer driver implementation. I too vote for controlling this by a board config option.
 
 > 
 > I would also recommend adding some sort of warning if the first
 > ioctl() is not available, to make potential regressions easier to
 > diagnose.
 > 

Thanks for pointing that out. Anyway if I go ahead with a board config option to select the ioctl to be used, based on vendor FB implementation, then it does't make sense to fall back to the other ioctl. Does it?
 
[1] https://groups.google.com/forum/#!topic/android-framework/2-S9ialdeNs

[2] https://groups.google.com/d/msg/android-porting/6Tn36GpVWJI/UApQ7C3eIEkJ

Patch-set: 1
Label: Verified=0