Update patch set 1
Patch Set 1:

Yes, no reason to print a message the fallback if we control which ioctl through a config option :-) Generally speaking, when porting the system to a new platform, there are always unexpected things happening at the kernel/guest interface, and having log messages to indicate this helps everyone tremendously. Thanks for your understanding!

Patch-set: 1