Update patch set 2
Patch Set 2:

One open question is the naming style of the new functions. When I wrote the first patch, I following the style of the existing registerBuffer() and unregisterBuffer() functions, which already conflict with the naming of the lock_ycbcr() function.

Now we have a new function called lockAsync_ycbcr(), an odd mix of both camelCase and underscore_case.

Maybe Jesse or Greg can advise on Google's preference?

Patch-set: 2