gralloc: Add copy buffer API
The new gralloc copyBuffer() method provides a way to copy the contents
from one buffer to another. The gralloc module then has the oportunity
to implement this as a hardware accelerated copy operation, which should
be more efficient than the client doing a lock/unlock and CPU copy
between the buffers.

Change-Id: I15dbaa61c1f581ee595c219560f1f7480c6d5914