Update patch set 2
Patch Set 2:

We're worried that at some point we may have to support buffers whose contents could change without a buffer handle update, so a buffer handle check is probably less than ideal.

The reason I like the solution I proposed is that it doesn't require any "translation" for older code paths that don't provide damage. For example, eglSwapBuffers(...) will directly map to eglSwapBuffersWithDamage(..., 0, NULL) instead of having to query/track the dimensions of the buffer in order to create a full-buffer rect, and passing that notion through the system will simplify the plumbing a bit.

Patch-set: 2