Update patch set 2
Patch Set 2:

I think damage would be pre-rotation, like sourceCrop.

On that note, I realized that other rects/regions in this struct assume the origin is top-left, but the EGL call is going to define rects from bottom-left. We should pick a convention and codify it in a comment. My inclination is top-left for consistency with the rest of the struct, but I could probably be convinced otherwise.

Patch-set: 2
Label: Verified=0