Update patch set 1
Patch Set 1:

KM_INVALID_KEY_BLOB, probably. Non-secure world should never actually have access to any of these key blobs and won't be able to use them at all if it somehow does, not even to export a public key.

This is to allow bootloaders to use keymaster for verified boot crypto, confident that non-secure can not muck with it.

Patch-set: 1
Label: Verified=0