Update patch set 1
Patch Set 1: Code-Review-1

There is no plan for supporting volume buttons in the audio framework.
The jack detection you are referring to is actually the antenna detection. It happens that this coincides with jack detection on a smartphone but this is not necessary the case, thus the need to support it in the FM HAL and framework.
The case of volume buttons is very different in that they are already supported by the input framework and I don't see a need to add support in the FM radio framework.

Patch-set: 1
Reviewer: Eric Laurent <1003259@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1