Add system/lib/framework/hw to the module search paths list
Due to b/33681361, libraries that were installed to /system/lib/hw are
now installed to /system/lib/framework/hw. In order to reflect the
change, search path is here is also updated.

Note that this patch is temporary because HAL modules are now required
to be under /vendor/lib. When they are all moved to the new place (
b/34135607), this patch must be reverted.

Test: build & run

Change-Id: I854ed1fc91856eaa5027e618747c8520229ced64