Update patch set 2
Patch Set 2: Code-Review-2

This enum should have an unsigned underlying type, ideally uint32_t, though a compiler could (bizarrely, IMO) choose to use an int64_t, since that's also capable of containing all of the values in the enum.

A better approach is to specify the exact type, which can be done in C++11 (e.g. enum hw_authenticator_type_t : uint32_t {...), but not in C.

Given that we need to maintain compatibility with C, the right approach is to leave the definition as-is, with UINT32_MAX as the largest value and all values positive, and explicitly cast to uint32_t at the point of usage that's generating the error.

Patch-set: 2
Reviewer: Shawn Willden <1043663@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2