Update patch set 3
Patch Set 3:

From the C++ standard 7.2/7

"For an enumeration whose underlying type is not fixed, the underlying type is an integral type that can represent all the enumerator values defined in the enumeration. If no integral type can represent all the enumerator values, the enumeration is ill-formed. It is implementation-defined which integral type is used as the underlying type except that the underlying type shall not be larger than int unless the value of an enumerator cannot fit in an int or unsigned int."

One of the unfortunate ways in which the language definitions have diverged, apparently.

On platforms where int is 64 bits, your change would force the compiler to use a 64-bit value for the enum type, and on platforms where it's 32 bits, it would result in implementation-defined behavior.

This change is too risky. It could break existing code in subtle and hard-to-understand ways. Given the fact that it's also going away, I think it's better to leave it as-is. The new code that's replacing it is strictly C++, so we'll no longer have to deal with the C/C++ ambguity.

Patch-set: 3
Reviewer: Shawn Willden <1043663@85c56323-6fa9-3386-8a01-6480fb634889>