Update patch set 1
Patch Set 1:

Yes, you may have to abandon this CL to enable a build on the C++ switch to finish first :-(, then merge down and be cleaned.

Then this one can be re-activated after all is well.

You may never get a clean TreeHugger from the C++ switch here, if merge conflicts break it. You then Submit with a message to TreeHugger that you are taking responsibility for the merge conflicts and override it.

Yes, a different sequence could have been done, hindsight is 20/20, but once the first one that got rid of the compiler warning was submitted to internal, this all got messy. (consider that punishment for not going as far upstream as you can go ;-}, here in AOSP, first).

Patch-set: 1
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>