Update patch set 1
Patch Set 1:

Peng,

Thanks for reviewing this so promptly.

Our intention is for this to be generic for all Android devices, not specific to any vendor. With that, we don't think a build flag makes sense.

I am working on your other comments, and will upload the patch set shortly.

Thanks,
-Kim

Patch-set: 1
Reviewer: Kim Low <1152828@85c56323-6fa9-3386-8a01-6480fb634889>