Update patch set 4
Patch Set 4:

(1 comment)

For the refactoring, can we do it in two steps? First the functionality patch, and after it is merged, we submit a refactoring patch? Or would you prefer a single patch?

Thanks for the +1. Can you help us ping someone for +2?

Patch-set: 4
Reviewer: Kim Low <1152828@85c56323-6fa9-3386-8a01-6480fb634889>