Update patch set 3
Patch Set 3:

Casting in the four places this is required seems more reasonable than turning off the warning completely.

Once this is fixed we can move it over to Soong and get it in the continuous builds. It's not building in our continuous builds right now because it's hidden behind USE_CAMERA_V4L2_HAL.

This was hidden behind USE_CAMERA_V4L2_HAL because some devices shipped V4L2 headers that were too old or otherwise incompatible with this code. Soong defaults to using the bionic kernel headers, only using the device-specific ones when requested, so this should be safe to compile on all builds once it's moved.

Patch-set: 3
Reviewer: Dan Willemsen <1074628@85c56323-6fa9-3386-8a01-6480fb634889>