Update patch set 1
Patch Set 1:

A bit of background on this change...

I've been working on our gralloc1 implementation, and trying clean up some of the error log prints we get for usage flags. I noticed the mismatch for the GPU_DATA_BUFFER usage flag, looks like the driver level treats it as bit 23, but java apps would treat it as bit 24.

But, we don't actually ever use the GPU_DATA_BUFFER flag, so for us this error is benign.

Patch-set: 1
Reviewer: Francis Hart <1044843@85c56323-6fa9-3386-8a01-6480fb634889>