camera: add a lock for shared containers
Buffers and frames "in flights" are accessed from several threads.
There are at least three threads which access to those data (see
`v4l2_camera.cpp` for more details):
 - "main" thread;
 - separate buffer enqueue thread ("Enqueue buffers");
 - separate buffer dequeue thread ("Dequeue buffers").
It is proposed to introduce additional mutex for containers to
prevent a race condition. As far as the most frequent functions
(`Add` and `Remove`) modify both containers it is proposed use only
one mutex for both containers.

Test: try to use a camera with more than one buffer and get SIGSEGV
      crash of the `cameraserver` in the `RequestTracker` class.
Change-Id: I2c1901722289cee9f0cb05a11a4dabe70d18f2ba
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>