Update patch set 1
Patch Set 1: Code-Review-2

Looked into this a bit. Unfortunately breaking binary compatibility of HALs once we've shipped them is virtually impossible.

We worked around this in the default implementation of a.h.graphics.allocator by explicitly translating the bit from the a.h.graphics.allocator definition (24) to the gralloc1 definition (23). So in practice this mismatch shouldn't result in any observable bug -- but please let us know if that's not the case.

commit 2d02d91f7fcd14f7fc523627ecf7af0682406fd7
Author: Chia-I Wu <olv@google.com>
Date:   Mon Jun 12 15:14:37 2017 -0700

    graphics: fix gralloc1 usage flags in default impl
    
    BufferUsage::SENSOR_DIRECT_DATA should be filtered out for consumer
    usage.  BufferUsage::GPU_DATA_BUFFER should be filtered out for
    producer usage and be translated to
    GRALLOC1_CONSUMER_USAGE_GPU_DATA_BUFFER for consumer usage.
    
    Bug: 62491483
    Test: boots
    Change-Id: I171712adf45ae80dadeed7997fb9872ff176fa1a

Patch-set: 1
Reviewer: Gerrit User 1010672 <1010672@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2