Update patch set 1
Patch Set 1:

(3 comments)

I've added some comments on making sure things are atomic, but also I'm not convinced this is correct from a design perspective (but I'm open to arguments in favor). Having a data structure be thread-safe adds overhead to people using it in a non-threaded manner, so to me it makes more sense to introduce the locks/synchronization wherever the data structure is actually being used in a multithreaded manner (in this case, the Camera class).

Patch-set: 1