Update patch set 11
Patch Set 11:

Some concerns were raised internally about the approach for this feature (ref: b/111457577), primarily a method for apps to know that some dynamic sensors are associated with a particular game controller/input device (and not the phone/Android device itself). Overall, I don't think the timing is right for merging this as it is not an approved feature for Q, and we need to have some further design discussions with michaelwr@ in the loop.

Patch-set: 11
Reviewer: Gerrit User 1094493 <1094493@85c56323-6fa9-3386-8a01-6480fb634889>