Update patch set 11
Patch Set 11:

For Android Q a lot of DualShock 4 related code was added including Bluetooth fixes, kernel driver, input button/stick mapping and others. On the area of Sensors, we had a meeting in Mountain View with Input team (including Michael) and others. At that time we had 2 patch series for Motion Sensors: one using "input framework for sensors" and other "evdev through motion sensors framework". The consensus was that using Sensors framework was conceptually the right way to go.

Device association was deferred until later (it also affects audio and other frameworks), but it didn't affect the correctness of the evdev Sensor code itself. The sensor is a dynamic sensor and applications can find it. Since it is a dynamic sensor it should not be the "default" as reported by "SensorManager.getDefaultSensor".

The applications can get information about the sensor through "getName()" or "getVendor()". On Android Q (which has DualShock 4 driver) this now reports:
getVendor(): Bluetooth 054c:09cc
getName(): Wireless Controller Motion Sensors
This capability is sufficient for PlayStation apps we intend to roll out during Q time frame, which depend on sensors.

Patch-set: 11
Reviewer: Gerrit User 1154383 <1154383@85c56323-6fa9-3386-8a01-6480fb634889>