Update patch set 1
Patch Set 1:

Yes, our Trusty team might be using hikey devices for testing, but they don't feed back into the community effort for hikey, so that isn't a valid reason to block this CL. Hikey needs to have its shipping FCM level increased to 29, which requires a gatekeeper HAL. This should of course be hardware-backed, but it isn't. I think this is a small price to pay.

To discourage others from actually using this code, I made it a static library, called it "software" and not "default", and this CL and the comments here should help with that.

Patch-set: 1
Reviewer: Gerrit User 1293738 <1293738@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0