Replace audio.primary.default with AVD-compatible HAL
This is a copy of audio.primary.cutf (cuttlefish audio HAL)

The purpose of this HAL implementation is to serve as common audio HAL
implementation for virtual devices.  Requires PCM device be available;
however, it is compatible with many existing hypervisor implementations
/ real hardware.

The previous HAL is still useful as a stub implementation and for
devices without speakers (but with microphone) such as the VR device;
it is kept as audio.miconly.default.

Change-Id: I472df8af3ff66a01a79767c6bb7074e3f44e14f7