Update patch set 3
Patch Set 3: Code-Review-1

The point of the original code is to have a virtual device implementation that will work when there is no real hardware. This is a requirement.

This new code requires a PCM device, so I don't think this can be submitted as is.

Why change the original code again?

Patch-set: 3
Reviewer: Gerrit User 1122301 <1122301@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1