Update patch set 1
Patch Set 1: Code-Review+2

This change looks fine to me. 

Just as a heads up, we are requiring new launching devices ship with gralloc 4.0 (graphics IAllocator 4.0/IMapper 4.0) in Android R. We will be deprecating this very old version of gralloc as soon as we can. It is not secure and it is keeping us from supporting new features in the framework. We will keep it as long as upgrading devices are using it. After that, emulators will be forced to upgrade.

Patch-set: 1
Reviewer: Gerrit User 1073546 <1073546@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2