Update patch set 1
Patch Set 1:

(2 comments)

Hi Peter,

By way of a very brief introduction I work with KevinB@ARM doing some work at the intersection of GPUs, FVPs and Android. 

For background the reason the framebuffer starts of in a 16 bpp config. is the "max-bandwidth" DTS setting. See (for example)

https://android.googlesource.com/kernel/arm64/+/refs/tags/android-10.0.0_r0.13/drivers/video/fbdev/amba-clcd.c#679

Kevin has proposed a patch to remove this. 

I was initially under the impression that the amba-clcd driver would not let you change to a config that exceeded this, which led me here to comment on the patch, but on the assumption that it does indeed work for you (as in the framebuffer really does change to RGBX_8888) I went digging in the kernel driver and concluded that it does not seem to inspect that value again(!)

Nonetheless I made a couple of small observations while I was looking at the code that you might want to consider.

Robert

Patch-set: 1
CC: Gerrit User 1584038 <1584038@85c56323-6fa9-3386-8a01-6480fb634889>