Update patch set 2
Patch Set 2:

the bug says this is so that we fuzz the mp4 extractor. It doesn't feel right that we need libhardware
in order to do that. 

can you give me the chain of dependencies that leads us to need libhardware to build on the host
in order to fuzz the extractors.

maybe it's a false dependency (it got listed at one point but perhaps is no longer needed).

Patch-set: 2
Reviewer: Gerrit User 1108919 <1108919@85c56323-6fa9-3386-8a01-6480fb634889>