Add a function pointer to get active boot slot
bootcontrol 1.2 adds a new function to get the active boot slot
on the next boot. Since the getter is closely related to the gpt
implementation of bootcontrol, we add the header here and expose
to clients of libhardware. Devices with an old HAL should never
call this new getter function.

Bug: 173808057
Test: Vts tests for bootcontrol

Change-Id: I19135ed16bf0ba546e0a8d9cb2d521ec0a9c5bbc